+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.11.0] - 2022-03-24
+
 ### Fixed
 
 - In `impersonateUser` resolver, if `clId` is provided but `userId` is not, check CL document to see if userId exists there and then update storefront-permissions' user record